requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
prometheus-client==0.0.17
python-consul==0.6.1
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
